blank-area: object is a pointer event, parameters for sheet, region
The presentation type BLANK-AREA now has keyword (and thus
non-required) presentation type parameters SHEET for specifying a
sheet containing the blank area and REGION for specifying a
region (within the area of the sheet).

When the system constructs a presentation instance with presentation
type BLANK-AREA, the presentation object is a pointer event the sheet
and position of which are the sheet for which the blank area
presentation is constructed and the pointer position in coordinates
for that sheet. The SHEET presentation type parameter is also that
sheet and the REGION is a POINT the position of which is the pointer
position.

Subtyping of BLANK-AREA presentation types with parameters is based on
matching sheets, if supplied, and the region of the subtype being
contained in that of the supertype, if supplied.

Co-authored-by: Daniel Kochmański <daniel@turtleware.eu>