package, colors: define and export colors based on shared data
Instead of having a list of color names in package.lisp and a
separate (generated) list of color definitions in drawing/colors.lisp,
generate the data required for both into data/colors.sexp and process
it as appropriate via read/load-time evaluation.

For drawing/colors.lisp, this has the additional benefit of not
compiling (and producing compiler progress messages) for hundreds of
toplevel forms.