examples: demodemo: remove now-unused generic function DISPLAY
Previously, several demos added methods on the generic function which
introduced an ordering problem: on the one hand, the generic
function (in demodemo.lisp) should be defined before methods are
added. On the other hand, the demo application frames should be defined
before demodemo.lisp is loaded since the latter refers to the former.