recording: fix and complete "freezing" of stored line style
In addition to the line thickness, the dash pattern must also be
converted in order to not change when replayed. This change thus
freezes line thickness and line dashes and also fixes the
destructive modification of the medium line style introduced in the
previous incomplete change.

This change fixes and completes the change in 3ab5e83d.