tests: new utilities {WITH-COMPARISON,INVOKE-AND-COMPARE}-TO-REFERENCE
The new function and macro
{WITH-COMPARISON,INVOKE-AND-COMPARE}-TO-REFERENCE load a reference
image file from a data directory, execute a function to produce a new
image and compare the two.

On top of that, the new function IS-SAME-BITMAP can be used to perform
the comparison in the context of a test. In case of a failure,
IS-SAME-BITMAP prints a textual approximation of the expected image,
the actual image and the difference image.