tests: allow offsets in output comparison utilities
{INVOKE-AND-COMPARE,WITH-COMPARISON}-TO-REFERENCE accept OFFSET-{X,Y}
keyword arguments to specify the respective offset. This allows
comparing a translated but otherwise identical version of the output
to the reference image. This comparison mode is useful for ensuring
that output records render identically after being moved to a new
position.

The new macro WITH-COMPARISON-TO-REFERENCE* is like
WITH-COMPARISON-TO-REFERENCE but executes the body three times: 1)
just draw 2) record and replay 3) record, move output record,
replay. Each output is compared to the reference image and expected to
match it.