recording: fix clipping of DRAW-TEXT-OUTPUT-RECORD
DRAW-TEXT-OUTPUT-RECORD has GS-{TRANSFORMATION,CLIP}-MIXIN as its
superclasses such that GS-TRANSFORMATION-MIXIN is first in the class
precedence list. This lead to the :AROUND method on
REPLAY-OUTPUT-RECORD specialized to GS-TRANSFORMATION-MIXIN to be
executed first, establishing the medium transformation. The
REPLAY-OUTPUT-RECORD specialized to GS-CLIP-MIXIN then passed the
stored clipping region, which is represented in stream coordinates, as
the :CLIPPING-REGION drawing option which didn't work.

To fixed this problem, this change removes REPLAY-OUTPUT-RECORD method
specialized to GS-TRANSFORMATION-MIXIN. This means all other mixins
can assume the medium transformation to be the identity while the only
client of GS-TRANSFORMATION-MIXIN, DRAW-TEXT-OUTPUT-RECORD can apply
the transformation itself as needed.

In contrast, the (setf output-record-position) method that was
specialized to DRAW-TEXT-OUTPUT-RECORD is now specialized to
GS-TRANSFORMATION-MIXIN since its behavior seems universal.

A new test checks the correctness of the clipping region
application. The test has to be disabled for now since it triggers an
off-by-one error in the render backend.

Reported-by: Daniel Kochmański <daniel@turtleware.eu>