windowing: events: POINTER-MOTION-EVENT is not a POINTER-BUTTON-EVENT
The specification says

  The instantiable class that corresponds to any sort of pointer
  motion event. This is a subclass of POINTER-EVENT.

Commit 2cfc2341 moved the BUTTON slot from the POINTER-EVENT to the
POINTER-BUTTON-EVENT, effectively removing the slot from
POINTER-MOTION-EVENT and defining the new POINTER-SCROLL-EVENT to not
have the slot.

After that, commit 6e0c7bea changed POINTER-SCROLL-EVENT and
POINTER-MOTION-EVENT to be subclasses of POINTER-BUTTON-EVENT because
1) several places in the codebase pass the :BUTTON initarg when
creating POINTER-MOTION-EVENT instances 2) the information in the
BUTTON slot could be useful for non-POINTER-BUTTON-EVENT events.

1) can be fixed and 2) suggests that the /button state/ (like
EVENT-MODIFIER-STATE), not a button press/release /event/ could be
useful in non POINTER-BUTTON-EVENT events.

So this change changes the superclass of POINTER-MOTION-EVENT back to
POINTER-EVENT. POINTER-SCROLL-EVENTS remains a POINTER-BUTTON-EVENT,
though.