presentations: translators: match translator gesture type, allow overriding
The basic building block %TEST-PRESENTATION-TRANSLATOR now matches the
gesture type in addition to the other parts of the gesture. This
allows distinguishing pointer gestures such as press, release, click
and double click.

Some functions which call %TEST-PRESENTATION-TRANSLATOR are adjusted
to pass :override '(:type nil ...) to preserve the previous behavior,
while others allow the gesture type to be taken into account:

* FRAME-INPUT-CONTEXT-BUTTON-PRESS-HANDLER method uses
  FIND-INNERMOST-PRESENTATION-MATCH instead of
  FIND-INNERMOST-APPLICABLE-PRESENTATION

* FRAME-HIGHLIGHT-AT-POSITION,
  FRAME-COMPUTE-POINTER-DOCUMENTATION-STATE,
  TEST-PRESENTATION-TRANSLATOR, FIND-APPLICABLE-TRANSLATORS,
  PRESENTATION-MATCHES-CONTEXT-TYPE, CALL-PRESENTATION-MENU,
  FIND-INNERMOST-APPLICABLE-PRESENTATION specify
  :override '(:type nil ...) to preserve the previous behavior