frames: do not call GENERATE-PANES when changing the current layout
Before this change, changing the current layout caused the panes
defined in :PANES to be re-created and then arranged into the selected
layout.  After this change, the set of panes is not touched (unless
the application frame class is redefined) and the existing pane
instances are arranged into the selected layout instead.

To separate the actions of generating panes and arranging panes into
a layout, the latter is now performed in the new generic function
APPLY-LAYOUT. Therefore, DEFINE-APPLICATION-FRAMES also generates two
methods, one on GENERATE-PANES and one on APPLY-LAYOUT.

This change also makes layout changes and frame redefinition work
better with dependencies between forms in the :PANES option:

  (:panes (foo :label :label "foo")
          (bar :label :label "bar")
          (baz (vertically () foo bar)))

With such a specification of panes, the current layout is changed
without re-creating any of the panes and without disturbing the
parent-child relations between BAZ and FOO and BAZ and BAR.

For a frame redefinition, FOO and BAR will be re-initialized and BAZ
will be re-created. After that, the newly created or re-initialized
panes will be assembled into the current layout.